+++ domain-services/fiscal-event-post-processor/src/main/java/org/egov/models/FiscalEventLineItemUnbundled.java
@@ -49,11 +49,11 @@ public class FiscalEventLineItemUnbundled {
     @JsonProperty("referenceId")
     private String referenceId = null;
 
-    @JsonProperty("parentEventId")
-    private String parentEventId = null;
+    @JsonProperty("linkedEventId")
+    private String linkedEventId = null;
 
-    @JsonProperty("parentReferenceId")
-    private String parentReferenceId = null;
+    @JsonProperty("linkedReferenceId")
+    private String linkedReferenceId = null;
 
     @JsonProperty("amount")
     private BigDecimal amount = null;
@@ -69,6 +69,9 @@ public class FiscalEventLineItemUnbundled {
 
     @JsonProperty("auditDetails")
     private AuditDetails auditDetails = null;
+
+    @JsonProperty("attributes")
+    private Object attributes = null;
 
 }