+++ domain-services/fiscal-event-service/src/main/java/org/egov/migration/V20220224143800_MongoIndexScript.java
@@ -12,13 +12,13 @@ import org.bson.Document;
 
 @ChangeLog(order = "001")
 @Slf4j
-public class V20210921123800_MongoIndexScript {
+public class V20220224143800_MongoIndexScript {
     public static final String FE_INDEX_AUTHOR = "iFIX-Fiscal-Event-Service";
 
     /**
      * @param db
      */
-    @ChangeSet(id = "V20210914185000_createFiscalEventIndex", author = FE_INDEX_AUTHOR, order = "001")
+    @ChangeSet(id = "V20220224185000_createFiscalEventIndex", author = FE_INDEX_AUTHOR, order = "001")
     public void createFiscalEventIndex(MongoDatabase db) {
         MongoCollection<Document> mongoCollection = db.getCollection("fiscal_event");
         mongoCollection.createIndex(Indexes.ascending("eventType"));
@@ -26,17 +26,7 @@ public class V20210921123800_MongoIndexScript {
         mongoCollection.createIndex(Indexes.ascending("tenantId"));
         mongoCollection.createIndex(Indexes.ascending("referenceId"));
         mongoCollection.createIndex(Indexes.ascending("ingestionTime"));
-        mongoCollection.createIndex(Indexes.ascending("departmentEntity"));
-        mongoCollection.createIndex(Indexes.ascending("departmentEntity.hierarchyLevel"));
-        mongoCollection.createIndex(Indexes.ascending("departmentEntity.code"));
-        mongoCollection.createIndex(Indexes.ascending("departmentEntity.name"));
-        mongoCollection.createIndex(Indexes.ascending("departmentEntity.ancestry.id"));
-        mongoCollection.createIndex(Indexes.ascending("departmentEntity.ancestry.code"));
-        mongoCollection.createIndex(Indexes.ascending("departmentEntity.ancestry.name"));
-        mongoCollection.createIndex(Indexes.ascending("departmentEntity.ancestry.hierarchyLevel"));
-        mongoCollection.createIndex(Indexes.ascending("expenditure.id"));
-        mongoCollection.createIndex(Indexes.ascending("project.id"));
-        mongoCollection.createIndex(Indexes.ascending("parentEventId"));
+        mongoCollection.createIndex(Indexes.ascending("linkedEventId"));
         mongoCollection.createIndex(Indexes.ascending("amountDetails.coa.id"));
         mongoCollection.createIndex(Indexes.ascending("amountDetails.amount"));
     }