+++ domain-services/fiscal-event-service/src/main/java/org/egov/web/models/FiscalEventRequest.java
@@ -28,6 +28,5 @@ public class FiscalEventRequest {
     @JsonProperty("fiscalEvent")
     private List<FiscalEvent> fiscalEvent = null;
 
-
 }